+++ clients/go/ahp/reducers.go
@@ -499,7 +499,7 @@ func ApplyActionToSession(state *ahptypes.SessionState, action ahptypes.StateAct
 		state.Summary.Activity = a.Activity
 		return ReduceOutcomeApplied
 	case *ahptypes.SessionChangesetsChangedAction:
-		state.Summary.Changesets = append([]ahptypes.ChangesetSummary(nil), a.Changesets...)
+		state.Summary.Changesets = append([]ahptypes.Changeset(nil), a.Changesets...)
 		return ReduceOutcomeApplied
 	case *ahptypes.SessionConfigChangedAction:
 		if state.Config == nil {